requirements.txt
python-dotenv
openai
streamlit
fastapi
uvicorn
pydantic
python-multipart